// src/game/data/enums.ts
export const enum EStatType {
  NONE = 0,
  HP = 1,
  MP = 2,
  STR = 3,
  AGI = 4,
  INT = 5,
  CON = 6,
  STR_X = 7,
  AGI_X = 8,
  INT_X = 9,
  CON_X = 10,
  CRITICALHIT = 15,
  SPECIALTY = 16,
  OPPRESSION = 17,
  RAPIDITY = 18,
  ENDURANCE = 19,
  MASTERY = 20,
  CRITICALHIT_X = 21,
  SPECIALTY_X = 22,
  OPPRESSION_X = 23,
  RAPIDITY_X = 24,
  ENDURANCE_X = 25,
  MASTERY_X = 26,
  MAX_HP = 27,
  MAX_MP = 28,
  MAX_HP_X = 29,
  MAX_MP_X = 30,
  MAX_HP_X_X = 31,
  MAX_MP_X_X = 32,
  NORMAL_HP_RECOVERY = 33,
  COMBAT_HP_RECOVERY = 34,
  NORMAL_HP_RECOVERY_RATE = 35,
  COMBAT_HP_RECOVERY_RATE = 36,
  NORMAL_MP_RECOVERY = 37,
  COMBAT_MP_RECOVERY = 38,
  NORMAL_MP_RECOVERY_RATE = 39,
  COMBAT_MP_RECOVERY_RATE = 40,
  SELF_RECOVERY_RATE = 41,
  DRAIN_HP_DAM_RATE = 42,
  DRAIN_MP_DAM_RATE = 43,
  DAM_REFLECTION_RATE = 44,
  EVOLUTION_DAM_RATE = 45,
  STIGMA_POWER_RATE = 46,
  CHAR_ATTACK_DAM = 47,
  SKILL_EFFECT_DAM_ADDEND = 48,
  ATTACK_POWER_RATE = 49,
  SKILL_DAMAGE_RATE = 50,
  ATTACK_POWER_RATE_X = 51,
  SKILL_DAMAGE_RATE_X = 52,
  COOLDOWN_REDUCTION = 53,
  PARALYZATION_POINT_RATE = 54,
  DEF = 55,
  RES = 56,
  DEF_X = 57,
  RES_X = 58,
  DEF_X_X = 59,
  RES_X_X = 60,
  CRITICAL_PHYSICAL_INC_RATE = 65,
  CRITICAL_MAGICAL_INC_RATE = 66,
  DEF_PEN_RATE = 67,
  RES_PEN_RATE = 68,
  PHYSICAL_INC_RATE = 69,
  MAGICAL_INC_RATE = 70,
  SELF_SHIELD_RATE = 71,
  HIT_RATE = 72,
  DODGE_RATE = 73,
  CRITICAL_HIT_RATE = 74,
  CRITICAL_RES_RATE = 75,
  CRITICAL_DAM_RATE = 76,
  ATTACK_SPEED = 77,
  ATTACK_SPEED_RATE = 78,
  MOVE_SPEED = 79,
  MOVE_SPEED_RATE = 80,
  PROP_MOVE_SPEED = 81,
  PROP_MOVE_SPEED_RATE = 82,
  VEHICLE_MOVE_SPEED = 83,
  VEHICLE_MOVE_SPEED_RATE = 84,
  SHIP_MOVE_SPEED = 85,
  SHIP_MOVE_SPEED_RATE = 86,
  FIRE_DAM_RATE = 87,
  ICE_DAM_RATE = 88,
  ELECTRICITY_DAM_RATE = 89,
  EARTH_DAM_RATE = 91,
  DARK_DAM_RATE = 92,
  HOLY_DAM_RATE = 93,
  ELEMENTS_DAM_RATE = 94,
  FIRE_RES_RATE = 95,
  ICE_RES_RATE = 96,
  ELECTRICITY_RES_RATE = 97,
  EARTH_RES_RATE = 99,
  DARK_RES_RATE = 100,
  HOLY_RES_RATE = 101,
  ELEMENTS_RES_RATE = 102,
  ULTIMATE_AWAKENING_DAM_RATE = 104,
  SELF_CC_TIME_RATE = 105,
  ENEMY_CC_TIME_RATE = 106,
  IDENTITY_VALUE1 = 107,
  IDENTITY_VALUE2 = 108,
  IDENTITY_VALUE3 = 109,
  AWAKENING_DAM_RATE = 110,
  ITEM_DROP_RATE = 111,
  GOLD_RATE = 112,
  EXP_RATE = 113,
  ATTACK_POWER_ADDEND = 123,
  ATTACK_POWER_ADDEND_2 = 124,
  NPC_SPECIES_HUMANOID_DAM_RATE = 125,
  NPC_SPECIES_DEVIL_DAM_RATE = 126,
  NPC_SPECIES_SUBSTANCE_DAM_RATE = 127,
  NPC_SPECIES_UNDEAD_DAM_RATE = 128,
  NPC_SPECIES_PLANT_DAM_RATE = 129,
  NPC_SPECIES_INSECT_DAM_RATE = 130,
  NPC_SPECIES_SPIRIT_DAM_RATE = 131,
  NPC_SPECIES_WILD_BEAST_DAM_RATE = 132,
  NPC_SPECIES_MECHANIC_DAM_RATE = 133,
  NPC_SPECIES_ANCIENT_DAM_RATE = 134,
  NPC_SPECIES_GOD_DAM_RATE = 135,
  NPC_SPECIES_ARCHFIEND_DAM_RATE = 136,
  VITALITY = 137,
  SHIP_BOOTER_SPEED = 138,
  SHIP_WRECK_SPEED_RATE = 139,
  ISLAND_SPEED_RATE = 140,
  ATTACK_POWER_SUB_RATE_1 = 141,
  ATTACK_POWER_SUB_RATE_2 = 142,
  PHYSICAL_INC_SUB_RATE_1 = 143,
  PHYSICAL_INC_SUB_RATE_2 = 144,
  MAGICAL_INC_SUB_RATE_1 = 145,
  MAGICAL_INC_SUB_RATE_2 = 146,
  SKILL_DAMAGE_SUB_RATE_1 = 147,
  SKILL_DAMAGE_SUB_RATE_2 = 148,
  RESOURCE_RECOVERY_RATE = 149,
  BASE_DAMAGE_RATE = 150,
  WEAPON_DAM = 151,
  WEAPON_DAM_X = 152
}

export const enum ECombatEffectActorType {
  NONE = 0,
  SELF = 1,
  TARGET = 2,
  CASTER = 3
}

export const enum ECombatEffectConditionType {
  NONE = 0,
  CURRENT_SKILL = 1,
  HP_LESS = 2,
  HP_GREATER = 3,
  MP_LESS = 4,
  MP_GREATER = 5,
  NPC_GRADE_LESS = 6,
  NPC_GRADE_GREATER = 7,
  NPC_GRADE_EQUAL = 8,
  NPC_SPECIES = 9,
  NPC_ATTR = 10,
  ABNORMAL_MOVE = 11,
  ABNORMAL_STATUS = 12,
  ABNORMAL_MOVE_IMMUNE = 13,
  ABNORMAL_STATUS_IMMUNE = 14,
  ABNORMAL_MOVE_ALL = 15,
  PC = 16,
  SKILL_EFFECT_ID = 17,
  IDENTITY_STACK_COUNT = 18,
  STATUS_EFFECT_IMMUNETYPE = 19,
  ABNORMAL_NOT_MOVE = 20,
  TARGET_COUNT = 21,
  SKILL_IDENTITY_CATEGORY = 22,
  IDENTITY_ELEMENT_VALUE = 23,
  DIRECTIONAL_ATTACK = 24,
  CURRENT_SKILL_GROUP = 25,
  ABNORMAL_MOVE_STATUS_ALL = 26,
  IDENTITY_STANCE = 27,
  PC_SKILL = 28,
  SKILL_EFFECT_GROUP_SET = 29,
  NPC_ID = 30,
  IDENTITY_GAUGE0_VALUE_LESS = 31,
  PC_WITHOUT_ME = 32,
  NPC_SCALED_LEVEL_EQUAL = 33,
  NPC_SCALED_LEVEL_LESS = 34,
  NPC_SCALED_LEVEL_GREATER = 35,
  NOT_SKILL_EFFECT_ID = 36,
  ABNORMAL_MOVE_NOT_IMMUNE = 37,
  APPLY_TARGET_MARKING = 38,
  DAMAGE_ATTR = 39,
  IDENTITY_ELEMENT_VALUE_LESS = 40,
  COMMAND_SKILL_TYPE = 41,
  STATUS_EFFECT_ON = 42,
  IDENTITY_GAUGE1_VALUE_LESS = 43,
  HIT_RANGE_OVER = 44,
  HIT_RANGE_UNDER = 45,
  DIRECTIONAL_SKILL_EFFECT = 46,
  IDENTITY_GAUGE0_VALUE_GREATER = 47,
  ULTIMATE_SKILL_POINT_GREATER = 48,
  ULTIMATE_SKILL_POINT_LESS = 49
}

export const enum ECombatEffectActionType {
  NONE = 0,
  MODIFY_DAMAGE = 1,
  MODIFY_FINAL_DAMAGE = 2,
  MODIFY_CRITICAL_RATIO = 3,
  MODIFY_CRITICAL_MULTIPLIER = 4,
  MODIFY_PENETRATION = 5,
  MODIFY_PENETRATION_WHEN_CRITICAL = 6,
  EXEC_ACTIVE_EFFECT_WHEN_DAMAGE = 7,
  EXEC_ACTIVE_EFFECT_WHEN_CRITICAL = 8,
  EXEC_REACTIVE_EFFECT_WHEN_MISS = 9,
  EXEC_REACTIVE_EFFECT_WHEN_DAMAGE = 10,
  EXEC_REACTIVE_EFFECT_WHEN_CRITICAL = 11,
  EXEC_AFTER_EFFECT = 12,
  EXEC_AFTER_SKILL = 13,
  APPLY_HEAL = 14,
  MODIFY_REACTIVE_DAMAGE = 15,
  MODIFY_DAMAGE_SHIELD_MULTIPLIER = 16,
  EXEC_ACTIVE_EFFECT_WHEN_KILL = 17,
  EXEC_START_SKILL = 18,
  MODIFY_PENETRATION_ADDEND = 19,
  MODIFY_PENETRATION_ADDEND_WHEN_CRITICAL = 20,
  MODIFY_REACTIVE_CRITICAL_MULTIPLIER = 21,
  MODIFY_DAMAGE_WHEN_CRITICAL = 22,
  MODIFY_PARALYZATION_POINT = 23,
  EXEC_WHEN_COUNTER = 24,
  EXEC_WHEN_BE_KILLED = 25,
  MODIFY_EVOLUTION_DAMAGE_MULTIPLIER = 26,
  MODIFY_AWAKENING_DAMAGE = 27
}
// ...
export enum EAbnormalMoveGroupType {
  NONE = 0,
  SWAY = 1,
  DOWN = 2,
  AIR = 3,
  PARALYZATION = 4
}

export enum EAbnormalStatusType {
  NONE = 0,
  FREEZE = 1,
  STONE = 2,
  FEAR = 3,
  STUN = 4,
  POLYMORPH = 5,
  SILENCE = 6,
  EARTHQUAKE = 7,
  VERMIN = 8,
  ELECTROCUTION = 9,
  CONFINEMENT = 10,
  SLEEP = 11,
  PROVOKE = 12,
  MIND_CONTROL = 13,
  FALL_DOWN = 14,
  DROWN = 15
}